avance del manejo de datos
