+++ packages/hosting/src/pages/practices/_practiceId/module1/annexs/annex4/index.jsx
@@ -4,7 +4,6 @@ import Col from "antd/lib/col";
 import Row from "antd/lib/row";
 import { Space } from "antd";
 import styled from "styled-components";
-import { Sheet2Integration } from "./Sheet2";
 import { Sheet1Integration } from "./Sheet1";
 
 export const Annex4Integration = ({
@@ -21,24 +20,8 @@ export const Annex4Integration = ({
       <Col span={24}>
         <div className="item-sheet">
           <Space direction="vertical" style={{ width: "100%" }}>
-            <Title level={4}>Hoja 1</Title>
+            <Title level={4}>Hoja de Evaluación</Title>
             <Sheet1Integration
-              practice={practice}
-              user={user}
-              users={users}
-              practitioner={practitioner}
-              company={company}
-              annex4={annex4}
-              onSavePractice={onSavePractice}
-            />
-          </Space>
-        </div>
-      </Col>
-      <Col span={24}>
-        <div className="item-sheet">
-          <Space direction="vertical" style={{ width: "100%" }}>
-            <Title level={4}>Hoja 2</Title>
-            <Sheet2Integration
               practice={practice}
               user={user}
               users={users}
