+++ packages/hosting/src/pages/practices/_practiceId/module1/annexs/annex6/Sheet1.jsx
@@ -13,7 +13,8 @@ import { useForm } from "react-hook-form";
 import { ProfessionalCareer } from "../../../../../../data-list";
 import * as yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
-import dayjs from "dayjs";
+import { Alert } from "antd";
+import { Link } from "react-router-dom";
 
 export const Sheet1Integration = ({ practice, practitioner }) => {
   const onConfirmSheet1 = () =>
@@ -32,11 +33,6 @@ export const Sheet1Integration = ({ practice, practitioner }) => {
 };
 
 const Sheet1 = ({ onConfirmSheet1, practitioner, practice }) => {
-  const ProfessionalCareerValue = ProfessionalCareer.find(
-    (profession) =>
-      profession.value === practitioner?.practitionerData?.professionalCareer
-  )?.label;
-
   const schema = yup.object({
     socialReason: yup.string().required(),
     address: yup.string().required(),
@@ -54,123 +50,32 @@ const Sheet1 = ({ onConfirmSheet1, practitioner, practice }) => {
       <Form>
         <Row gutter={[16, 16]}>
           <Col span={24}>
-            <Title level={5}>I. DATOS GENERALES DEL PRACTICANTE:</Title>
-          </Col>
-          <Col span={24} md={7}>
-            <div>
-              <label>Apellidos y Nombres: </label>
-              <p>
-                {`${practitioner?.paternalSurname} ${practitioner?.maternalSurname} ${practitioner?.firstName}` ||
-                  "-"}
-              </p>
-            </div>
-          </Col>
-          <Col span={24} md={7}>
-            <div>
-              <label>Carrera Profesional: </label>
-              <p>{ProfessionalCareerValue || "-"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={10}>
-            <div>
-              <label>
-                Módulo (s) Técnico-Profesional(es) ejecutado del Plan
-                Curricular:{" "}
-              </label>
-              <p>{practice?.name || "-"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={8}>
-            <div>
-              <label>
-                Período de Estudios (Año de ingreso - Año de Egreso) :
-              </label>
-              <p>
-                {practitioner?.practitionerData?.entryYear || "-"} -{" "}
-                {practitioner?.practitionerData?.yearGraduation || "En curso"}
-              </p>
-            </div>
-          </Col>
-          <Col span={24} md={8}>
-            <div>
-              <label>Domicilio: </label>
-              <p>{"-"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={8}>
-            <div>
-              <label>Teléfonos (Domicilio - Personal): </label>
-              <p>{practitioner?.phoneNumber || "-"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={8}>
-            <div>
-              <label>Correo Electrónico: </label>
-              <p>{practitioner?.email || "-"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={8}>
-            <div>
-              <label>DNI: </label>
-              <p>{practitioner?.dni || "-"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={8}>
-            <div>
-              <label>Código de Matrícula: </label>
-              <p>{practitioner?.practitionerData?.tuitionId || "-"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={6}>
-            <div>
-              <label>Semestre: </label>
-              <p>{practitioner?.practitionerData?.semester || "Egresado"}</p>
-            </div>
-          </Col>
-          <Col span={24} md={6}>
-            <div>
-              <label>Turno: </label>
-              <p>
-                {practitioner?.practitionerData?.studentShift || "Egresado"}
-              </p>
-            </div>
-          </Col>
-          <Col span={24} md={6}>
-            <div>
-              <label>Período de Evaluación: </label>
-              <p>
-                {practice?.startDate
-                  ? dayjs(practice.startDate, "DD/MM/YYYY").format("DD/MM/YYYY")
-                  : "-"}{" "}
-                -{" "}
-                {practice?.endDate
-                  ? dayjs(practice.endDate, "DD/MM/YYYY").format("DD/MM/YYYY")
-                  : "-"}
-              </p>
-            </div>
-          </Col>
-          <Col span={24} md={6}>
-            <div>
-              <label>Total de Horas: </label>
-              <p>{practice?.hours || "-"}</p>
-            </div>
+            <Title level={5}>
+              I. INFORMACIÓN GENERAL SOBRE EL PRACTICANTE:
+            </Title>
+            <Alert
+              type="info"
+              showIcon
+              message={
+                <>
+                  Antes de aprobar, revisa el anexo haciendo{" "}
+                  <a href={`/practices/${practice.id}/module1/sheets#annex6`}>
+                    Click Aquí!
+                  </a>
+                </>
+              }
+            />
           </Col>
         </Row>
-        <Row justify="end" gutter={[16, 16]}>
-          <Col span={24} sm={6} md={4}>
+        <Row justify="center" gutter={[16, 16]}>
+          <Col>
             <Button
               type="primary"
               danger
               size="large"
-              block
               onClick={() => onConfirmSheet1()}
             >
               Aprobar
-            </Button>
-          </Col>
-          <Col span={24} sm={6} md={4}>
-            <Button type="primary" size="large" block htmlType="submit">
-              Guardar
             </Button>
           </Col>
         </Row>
